The purpose of this milestone is to add an interactive map your FavouritePlaces project to have an app that uses an online map services.
Make sure you not only store a location name for each place in your app, but also its geographical coordinates (as already shown in the milestone 1 video).
Add a navigation link to your detail view to display the full information of that location, which includes the name of the location, its coordinates, as well as an interactive map showing the location.

Make sure that the map updates whenever the coordinates of the location change, and make sure that you allow the user to change the location (coordinates) interactively through the map.

Pan the map